+++ CHANGES.md
@@ -6,6 +6,7 @@ v0.2.3 (dev)
   with this code, I can now evaluate relatedness for 3756 in under 30 seconds on my laptop.
 + better scaling of X and Y depth
 + use final RG as the sample id in relate
++ output expected relatedness in .pairs.tsv file
 
 v0.2.2
 ======